% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_plausible_models.R
\name{plausible_models}
\alias{plausible_models}
\title{Select plausible models based on AIC and average stability}
\usage{
plausible_models(forest, pi, Delta = 2, tau = 0.4, jaccard_thresh = NULL)
}
\arguments{
\item{forest}{output from build_paths()}

\item{pi}{named numeric vector of variable stability}

\item{Delta}{AIC tolerance (default 2)}

\item{tau}{minimum average stability threshold (default 0.6)}

\item{jaccard_thresh}{optional, drop near-duplicate models with Jaccard similarity >= threshold}
}
\value{
data.frame of plausible models with columns: key, aic, size, vars, avg_stability
}
\description{
Select plausible models based on AIC and average stability
}
