% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{stability}
\alias{stability}
\title{Stability estimation via resampling and multi-path search}
\usage{
stability(
  x,
  y,
  B = 50,
  resample = c("bootstrap", "subsample"),
  m = NULL,
  build_args = list(),
  seed = NULL,
  verbose = FALSE
)
}
\arguments{
\item{x}{A \code{data.frame} (matrix) of numerical predictors.}

\item{y}{A \code{vector} of numerical or factor response.}

\item{B}{A \code{numeric} (integer) used to denote number of resamples (default = 50).}

\item{resample}{A \code{string} used to denote type of resampling. Must be bootstrap or subsample (default = "bootstrap").}

\item{m}{A \code{numeric} used to denote subsample size if resample="subsample" (default = floor(0.7*n)).}

\item{build_args}{\code{list} of arguments to pass to build_paths()(except x,y).}

\item{seed}{\code{integer} seed for reproducibility.}

\item{verbose}{A \code{logical} that determines whether the function prints progress reports as it runs (default = FALSE).}
}
\value{
A \code{list} with elements:
\describe{
\item{pi}{each variables stability}
\item{metadata}{parameters and counts}
}
}
\description{
Run multi-path forward selection on resampled datasets and computes
stability scores pi_j = average proportion of models containing predictor j across resamples.
}
\examples{
\dontrun{
stable <- stability(x = X, y = y, B = 50, resample = "bootstrap")
stable <- stability(x = X, y = y, B = 40, resample = "subsample", m = 100, verbose = TRUE)
}
}
\author{
Lijuan Wang, Kira Noordwijk, Evan Jerome
}
