% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_models.R
\name{plausible_models}
\alias{plausible_models}
\title{Select plausible models based on AIC and average stability}
\usage{
plausible_models(forest, pi, Delta = 2, tau = 0.4, jaccard_thresh = NULL)
}
\arguments{
\item{forest}{A \code{list} that is an output from build_paths()}

\item{pi}{A numeric \code{vector} of variable stability}

\item{Delta}{A \code{numeric} that denotes AIC tolerance for keeping near-best models (default = 2)}

\item{tau}{A \code{numeric} that denotes minimum average stability threshold (default = 0.4)}

\item{jaccard_thresh}{A \code{numeric} that denotes the optional threshold (0-1) to drop near-duplicate models(default = NULL).}
}
\value{
A \code{data.frame} of plausible models with columns: key, aic, size, vars, avg_stability
}
\description{
Filter models from a multi-path forward selection forest based on AIC,
average stability of predictors, and optional Jaccard similarity to remove near-duplicates.
}
\examples{
\dontrun{
plaus <- plausible_models(forest = forest, pi = pi)
plaus <- plausible_models(forest = forest, pi = pi, Delta = 3, tau = 0.5, jaccard_thresh = 1)
}
}
\author{
Lijuan Wang, Kira Noordwijk, Evan Jerome
}
