% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_stability.R
\name{stability}
\alias{stability}
\title{Compute variable stability via resampling}
\usage{
stability(x, y, B = 50, resample = "bootstrap", m = NULL, ...)
}
\arguments{
\item{x}{predictor matrix or data.frame}

\item{y}{response vector}

\item{B}{number of resamples}

\item{resample}{"bootstrap" or "subsample"}

\item{m}{subsample size if resample="subsample"}

\item{...}{additional parameters passed to build_paths()}
}
\value{
list with pi (variable stability) and metadata
}
\description{
Compute variable stability via resampling
}
